refactor(sms): improve sms sending logic
- Extracted sms logic into actions
- Improved phone number handling
- Reduced code duplication
- Enhanced error handling
- Improved UI responsiveness
