+++ app/Http/Controllers/NotificationController.php
@@ -1,10 +1,16 @@
 <?php
 namespace App\Http\Controllers;
 
+use App\Actions\SMS\AllDueStudentNumbers;
+use App\Actions\SMS\AllStaffNumbers;
+use App\Actions\SMS\AllStudentWithGuardianNumbers;
+use App\Actions\SMS\SelectedDepartmentClassesStudentNumbers;
+use App\Actions\SMS\SelectedStudentNumbers;
+use App\Actions\SMS\SMSAction;
+use App\Actions\SMS\SpecificPhoneNumbers;
 use App\Models\Department;
 use App\Models\NotificationToken;
 use App\Models\SMSBalance;
-use App\Models\User;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
@@ -117,7 +123,7 @@ class NotificationController extends Controller
     {
         // Validate the request
         $request->validate([
-            'sms_message'           => 'required|string|max:160',
+            'sms_message'           => 'required|string',
             'all_students'          => 'boolean',
             'all_staff'             => 'boolean',
             'due_students'          => 'boolean',
@@ -152,46 +158,17 @@ class NotificationController extends Controller
 
         // 1. All Students
         if ($allStudents) {
-            $students = User::whereHas('roles', function ($q) {
-                $q->where('name', 'student');
-            })->with('guardians')->get();
-
-            $phoneNumbers = $phoneNumbers->merge($this->extractPhoneNumbers($students));
+            $phoneNumbers = $phoneNumbers->merge(AllStudentWithGuardianNumbers::run());
         }
 
         // 2. All Staff
         if ($allStaff) {
-            $staff = User::whereHas('roles', function ($q) {
-                $q->where('name', 'staff');
-            })->get();
-
-            $phoneNumbers = $phoneNumbers->merge($this->extractPhoneNumbers($staff));
+            $phoneNumbers = $phoneNumbers->merge(AllStaffNumbers::run());
         }
 
         // 3. Due Students
         if ($allDueStudents) {
-            $currentDate = date('Y-m');
-            $dueStudents = User::with(['academics.class', 'academics.department', 'incomeLogs.studentDue', 'guardians'])
-                ->whereHas('roles', fn($q) => $q->where('name', 'student'))
-                ->where(function ($query) use ($currentDate) {
-                    // Students who don't have income log for current month
-                    $query->whereDoesntHave('incomeLogs', function ($q) use ($currentDate) {
-                        $q->where('payment_period', $currentDate)
-                            ->whereHas('feeType', function ($feeTypeQuery) {
-                                $feeTypeQuery->where('slug', 'like', '%academic-fee%')
-                                    ->orWhere('slug', 'like', '%boarding-fee%');
-                            });
-                    })
-                    // Or students who have a due for current month
-                        ->orWhereHas('incomeLogs', function ($q) use ($currentDate) {
-                            $q->where('payment_period', $currentDate)
-                                ->whereHas('studentDue', function ($dueQuery) {
-                                    $dueQuery->where('due_amount', '>', 0);
-                                });
-                        });
-                })->get();
-
-            $phoneNumbers = $phoneNumbers->merge($this->extractPhoneNumbers($dueStudents));
+            $phoneNumbers = $phoneNumbers->merge(AllDueStudentNumbers::run());
         }
 
         // 4. Selected Departments
@@ -199,43 +176,21 @@ class NotificationController extends Controller
             foreach ($departmentSelections as $selection) {
                 $departmentId  = $selection['departmentId'] ?? null;
                 $selectedClass = $selection['selectedClass'] ?? 'all';
-
                 if ($departmentId) {
-                    $query = User::whereHas('roles', function ($q) {
-                        $q->where('name', 'student');
-                    })->whereHas('academics', function ($q) use ($departmentId) {
-                        $q->where('department_id', $departmentId);
-                    })->with('guardians');
-
-                    // Filter by class if specific class is selected
-                    if ($selectedClass !== 'all') {
-                        $query->whereHas('academics', function ($q) use ($selectedClass) {
-                            $q->where('class_id', $selectedClass);
-                        });
-                    }
-
-                    $students     = $query->get();
-                    $phoneNumbers = $phoneNumbers->merge($this->extractPhoneNumbers($students));
+                    $phoneNumbers = $phoneNumbers->merge(SelectedDepartmentClassesStudentNumbers::run($departmentId, $selectedClass));
                 }
             }
         }
 
         // 5. Selected Student IDs
         if (! empty($selectedStudentIds)) {
-            $students = User::whereIn('unique_id', $selectedStudentIds)
-                ->with('guardians')->get();
-            $phoneNumbers = $phoneNumbers->merge($this->extractPhoneNumbers($students));
+            $phoneNumbers = $phoneNumbers->merge(SelectedStudentNumbers::run($selectedStudentIds));
         }
 
         // 6. Extra Numbers
         if (! empty($extraNumbers)) {
-            $extraNumbers = explode(',', $extraNumbers);
-            $extraNumbers = collect($extraNumbers)->map(function ($number) {
-                return validateAndFormatPhoneNumber(trim($number));
-            })->filter();
-            $phoneNumbers = $phoneNumbers->merge($extraNumbers);
+            $phoneNumbers = $phoneNumbers->merge(SpecificPhoneNumbers::run($extraNumbers));
         }
-
         // Remove duplicates and filter out empty numbers
         $uniquePhoneNumbers = $phoneNumbers
             ->filter(function ($number) {
@@ -244,86 +199,27 @@ class NotificationController extends Controller
             ->unique()
             ->values();
 
-        // Validate and format phone numbers
-        $validPhoneNumbers = $uniquePhoneNumbers->map(function ($number) {
-            return validateAndFormatPhoneNumber($number);
-        })->filter();
-
-        // all valid phone numbers separated by comma as array
-        $validPhoneNumbersAsArray = $validPhoneNumbers->implode(',');
         // calculate message character wise cost and total number of messages cost so 140 characters sms .50 tk
         $messageLength = strlen($smsMessage);
         if ($messageLength < 140) {
             $messageLength = 140; // Adjust length to 160 for cost calculation
         }
         $totalMessagesPart = ceil($messageLength / 140);
-        $totalCost         = $totalMessagesPart * $validPhoneNumbers->count() * 0.45; // Assuming 0.50 tk per 160 characters
+        $totalCost         = $totalMessagesPart * $uniquePhoneNumbers->count() * 0.45; // Assuming 0.50 tk per 160 characters
 
         $smsBalance = SMSBalance::first();
         if (! $smsBalance || $smsBalance->balance < $totalCost) {
             return redirect()->back()->with('error', 'Insufficient SMS balance. Please recharge your balance. Current balance: ' . $smsBalance->balance . ' tk' . ' Required balance: ' . $totalCost . ' tk');
         } else {
-            $res = $this->sms_send($validPhoneNumbersAsArray, $smsMessage);
+            $res = SMSAction::run($uniquePhoneNumbers, $smsMessage);
             $res = json_decode($res, true);
             if ($res['success_message']) {
                 $smsBalance->balanceDecrement($totalCost);
-                return redirect()->back()->with('success', 'SMS sent successfully to ' . $validPhoneNumbers->count() . ' recipients.');
+                return redirect()->back()->with('success', 'SMS sent successfully to ' . $uniquePhoneNumbers->count() . ' recipients.');
             } else {
                 return redirect()->back()->with('error', $res['error_message']);
             }
         }
-
-        // return redirect()->back()->with('success', "SMS prepared and queued successfully");
-    }
-
-    private function extractPhoneNumbers($users)
-    {
-        $phoneNumbers = collect();
-        foreach ($users as $user) {
-            // Get user's own phone number
-            if ($user->phone) {
-                $phoneNumbers->push($user->phone);
-            }
-
-            // Get guardian phone numbers
-            if ($user->hasRole('student') && $user->guardians && $user->guardians->numbers) {
-                $guardianNumbers = json_decode($user->guardians->numbers, true);
-                if (is_array($guardianNumbers)) {
-                    foreach ($guardianNumbers as $number) {
-                        if ($number) {
-                            $phoneNumbers->push($number);
-                        }
-                    }
-                }
-            }
-        }
-
-        return $phoneNumbers;
-    }
-
-    public function sms_send($validPhoneNumbers, $message)
-    {
-        $url      = "http://bulksmsbd.net/api/smsapi";
-        $api_key  = "WqEMXtKIyUzdOiLrQyGm";
-        $senderid = "8809617622335";
-        $number   = $validPhoneNumbers;
-
-        $data = [
-            "api_key"  => $api_key,
-            "senderid" => $senderid,
-            "number"   => $number,
-            "message"  => $message,
-        ];
-        $ch = curl_init();
-        curl_setopt($ch, CURLOPT_URL, $url);
-        curl_setopt($ch, CURLOPT_POST, 1);
-        curl_setopt($ch, CURLOPT_POSTFIELDS, $data);
-        curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-        curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false);
-        $response = curl_exec($ch);
-        curl_close($ch);
-        return $response;
-
     }
 
 }
